Fix the areatrigger table not being filled and also fix the types related to the areatriggers

+++ SAI-Editor/Classes/Database/Classes/AreaTrigger.cs
@@ -7,9 +7,9 @@
         public double posX { get; set; }
         public double posY { get; set; }
         public double posZ { get; set; }
-        public double field5 { get; set; } // unk
-        public double field6 { get; set; }
-        public double field7 { get; set; }
+        public int field5 { get; set; } // unk
+        public int field6 { get; set; }
+        public int field7 { get; set; }
         public double field8 { get; set; }
         public double field9 { get; set; }
     }
